Errors handling is exposed
